opendata/requirements.txt
pandas==0.23.4
numpy==1.15.3
yamlordereddictloader==0.4.0
PyYAML==3.13
ddlparse==1.1.1
jours-feries-france==0.3.0
vacances-scolaires-france==0.4
nose==1.3.7
